refactor: dissolve other_tools.py into proper category files
